feat: relocate wagmi config
