+++ frontend/pages/login/components/EthereumSignIn.tsx
@@ -1,7 +1,7 @@
 import { useAccount, useNetwork } from "wagmi";
 import ConnectButton from "./ConnectButton";
 import AddressPill from "./AddressPill";
-import { isChainIdSupported } from "@/wagmi.config";
+import { isChainIdSupported } from "@/config/wagmi.config";
 import { CustomButton } from "@components/Button";
 import { useTranslation } from "react-i18next";
 import LoginButton from "./LoginButton";
