+++ frontend/pages/login/components/LoginButton.tsx
@@ -1,5 +1,5 @@
 import { useSiweIdentity } from "@/siwe";
-import { isChainIdSupported } from "@/wagmi.config";
+import { isChainIdSupported } from "@/config/wagmi.config";
 import { CustomButton } from "@components/Button";
 import { useAccount, useNetwork } from "wagmi";
